AP-4324: Extract Individual model from Opponent
Extract from opponent with polymorphic relationship to
opponennt.
